Pong's fundamentals in verilog
In these files, i coded the very fundamentals of the game of pong, such as collisions, the movement of the players, etc etc. Also, it's noteworthy that the software im using is Quartus Prime 21.1, and the pins that are set for this project are specific for the board im using, and may be different for the board you, dear reader, may be using. 
Right now, i am sort of in a rush, so i will only upload the files, but i will attempt to go into further explanation of the code in the future, maybe, but there definitely will be something explaining at least ... well.,... something in this code. Oh, and the explanation on configuring VGA in the altera DE2-115 board, yeah, i will just post that on the readme file i think